+++ Plugins/Reqnroll.Microsoft.Extensions.DependencyInjection.ReqnrollPlugin/MissingScenarioDependenciesException.cs
@@ -6,7 +6,7 @@ namespace Reqnroll.Microsoft.Extensions.DependencyInjection
     public class MissingScenarioDependenciesException : ReqnrollException
     {
         public MissingScenarioDependenciesException()
-            : base("No method marked with [ScenarioDependencies] attribute found.")
+            : base("No method marked with [ScenarioDependencies] attribute found, or unsupported return type specified.")
         {
             HelpLink = "https://go.reqnroll.net/doc-msdi";
         }